requirements.txt
tensorflow
numpy
matplotlib
seaborn
scikit-learn
opencv-python
Pillow
h5py
